[#965] Fix for type errors in composable
